ghost_example/requirements.txt
aws-cdk.core
aws-cdk.aws_ec2
aws-cdk.aws_rds
aws-cdk.aws_eks
aws-cdk.aws_iam
pyyaml
